Benford's Law
The thing is, Benford's Law doesn't apply when you have a specific range of random data to extract from.
The maximum number of the data set that you are extracting random information from must be undetermined, or infinite.
In other words, say, you used a computer number generator that had a 'set' or specific range from which to extract
the numbers, eg. 1-100. You would undoubtedly end up with a random dataset of numbers, yes, but the number 1 would
appear as a first digit as often as the number 9 or any other number.

**The interesting** part, actually, happens when you let a computer (or nature) decide randomly, and on each instance,
how large you want the random number to potentially be. Then you get a nice, bi-dimensional random dataset, that perfectly
attains to Benford's Law. I have generated this RUBY code, which will neatly prove that, to every Mathematician's fascination,
Benford's Law works each and every single time for randomly chosen numbers, of randomly decided length!